Update and rename db-koth.js to DBKoth.js
Adds DB sync every 60 seconds
Adds syncing files to the playerdata folder if they do not exist already from the DB directly. For example a new server is deployed it can sync with the DB and pull all the players data before any player has a chance to join it.